increase gles shader precision

+++ system/shaders/GLES/2.0/gles_shader.vert
@@ -26,7 +26,7 @@ attribute vec4 m_attrcord0;
 attribute vec4 m_attrcord1;
 varying vec4 m_cord0;
 varying vec4 m_cord1;
-varying lowp vec4 m_colour;
+varying highp vec4 m_colour;
 uniform mat4 m_proj;
 uniform mat4 m_model;
 uniform mat4 m_coord0Matrix;
