+++ system/shaders/GLES/2.0/gles_shader_fonts.frag
@@ -23,7 +23,7 @@
 precision mediump float;
 uniform sampler2D m_samp0;
 varying vec4 m_cord0;
-varying lowp vec4 m_colour;
+varying highp vec4 m_colour;
 
 void main ()
 {
