+++ system/shaders/GLES/2.0/gles_shader_rgba_bob_oes.frag
@@ -27,7 +27,7 @@ uniform samplerExternalOES m_samp0;
 uniform samplerExternalOES m_samp1;
 varying vec4 m_cord0;
 varying vec4 m_cord1;
-varying lowp vec4 m_colour;
+varying highp vec4 m_colour;
 uniform int m_method;
 uniform int m_field;
 uniform float m_step;
